feat(super luck): use updated probabilities
New data from RaenonX has more reliable estimates for the probabilities
of each outcome for this skill. RaenonX is using different
probabilities for each outcome, so I've rewritten the skill to support
arbitrary probability distributions for the outcomes.

The current distribution is based on a best-fit I ran with the
following assumptions:

* The ingredients don't need to have the same likelihoods as each
  other.
* Each probability is a multiple of 1%.