fix(tierlist): Re-bucket tiers
With the change to add dish strength to scores, all mons have significantly higher scores, so more things are within 90% of the top. If we change the thresholds for each tier, then we'll probably have this problem come up again as newer and stronger dishes are added. Instead, we can subtract out the bottom score from everything, so that the lowest score is 0, guaranteeing we'll always have an F tier, rather than having everything bunch up in the middle tiers.