refactor(pokemon): use util funcs to make Pokemon
Everywhere we define an entire Pokemon all at once (rather than describing one in terms of another, as we do for evolutions), add in a call to one of the new util functions.

Checked for missing mons by adding a mock field to Pokemon and looking for type errors within common.