fix(berry-burst): remove range from skill detail
The Berry Burst skill detail vue file was copied from the file for Disguise. With Disguise, we show that each activation produces x-y berries, where x is the non-crit amount and y is the crit amount. For regular Berry Burst, there's no crit, but we still show a range of x-x. Change that to just show one number.