refactor: add constructors for evolved pokemon
We already have util functions for copying most of the fields from a base mon. An advantage to the new functions is requiring that certain fields be overwritten when defining the new mon. Previously, if Weavile's frequency was left unspecified, it would use Snealsel's frequency. Now, it would instaed be a compilation error.