feat(amber): add new mons
* Trapinch: https://pks.raenonx.cc/en/pokedex/328
* Vibrava: https://pks.raenonx.cc/en/pokedex/329
* Flygon: https://pks.raenonx.cc/en/pokedex/330
* Bagon: https://pks.raenonx.cc/en/pokedex/371
* Shelgon: https://pks.raenonx.cc/en/pokedex/372
  * ing% and skill% not yet settled
* Salamence: https://pks.raenonx.cc/en/pokedex/373
* Dwebble: https://pks.raenonx.cc/en/pokedex/557
* Crustle: https://pks.raenonx.cc/en/pokedex/558

Ingredient Draw (without modifiers) is finally a skill, and the ingredients match what can be found by the mons that get it. For now, I'm assuming that's always how it works. Testing with Mr. Mime and Togekiss might reveal that it's always Potato, Oil, and Avocado.

I'm not sure how to indicate that the skill _can_ be targetted by Metronome, but maybe it's okay to skip that until we get confirmation that it can be targetted.

I've heard the RP project's data is no longer publicly available, so I don't know the RP values for Ingredient Draw. But I'm going to assume they're the same as Hyper Cutter and Super Luck, until I learn otherwise.